@@ -24,20 +24,20 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@angular/common": "~4.3.4",
-    "@angular/compiler": "~4.3.4",
-    "@angular/core": "~4.3.4",
-    "@angular/forms": "~4.3.4",
-    "@angular/http": "~4.3.4",
-    "@angular/platform-browser": "~4.3.4",
-    "@angular/platform-browser-dynamic": "~4.3.4",
-    "@angular/router": "~4.3.4",
+    "@angular/common": "~5.0.1",
+    "@angular/compiler": "~5.0.1",
+    "@angular/core": "~5.0.1",
+    "@angular/forms": "~5.0.1",
+    "@angular/http": "~5.0.1",
+    "@angular/platform-browser": "~5.0.1",
+    "@angular/platform-browser-dynamic": "~5.0.1",
+    "@angular/router": "~5.0.1",
 
-    "angular-in-memory-web-api": "~0.3.0",
-    "systemjs": "0.19.40",
-    "core-js": "^2.4.1",
-    "rxjs": "5.0.1",
-    "zone.js": "^0.8.4"
+    "angular-in-memory-web-api": "~0.5.1",
+    "systemjs": "0.20.19",
+    "core-js": "^2.5.1",
+    "rxjs": "5.5.2",
+    "zone.js": "^0.8.18"
   },
   "devDependencies": {
     "concurrently": "^3.2.0",
